requirements.txt
numpy
opencv-python
mediapipe
opencv-python 
paho-mqtt
kubernetes
termcolor